feat: Support dynamic port name in ocean chart (#232)
* Add portName configuration to services in values.yaml

Updated the values.yaml file to include a portName for both the main service and live events service. Adjusted service templates to utilize the portName value, defaulting to 'ocean-port' if not specified.

* Update service template to reference portName from values.yaml

Modified the service.yaml template to dynamically use the portName defined in values.yaml, defaulting to 'ocean-port'. This change enhances configurability for the actions processor service.

* bump chart version

* Revert chart version to 0.11.0